NO-TICKET: add component wrapper
