+++ clients/mobile/src/utility/deviceRegistration.ts
@@ -1,9 +1,5 @@
-import { useFocusEffect } from '@react-navigation/native'
-import { useCallback } from 'react'
-import { InteractionManager } from 'react-native'
 import authService from '../authentication/services/authService'
 import type { HttpClient } from '../setup/clients/HttpClient'
-import { useFetchClient } from '../setup/clients/useFetchClient'
 import { TOKEN } from '../setup/constant/token'
 import {
   saveDeviceTokenOnSNS
@@ -13,22 +9,21 @@ import {
 } from '../setup/notification/registerForPushNotifications'
 import StorageService from './storageService'
 
-const registerUserDeviceForNotification = (fetchClient: HttpClient): void => {
+const registerUserDeviceForNotification = (fetchClient: HttpClient): Promise<void> =>
   registerForPushNotificationsAsync().then(async token => {
     const loggedInUser = await authService.currentLoggedInUser()
 
     if (!await isDeviceAlreadyRegisteredForUser(token, loggedInUser.userId)) {
       await saveDeviceTokenOnSNS(token as string, fetchClient)
     }
+  }).catch(error => {
+    throw new Error(
+      `Failed to register user device for notifications: ${
+        error instanceof Error ? error.message : 'An unexpected error occurred'
+      }`
+    )
   })
-    .catch(error => {
-      throw new Error(
-        `Failed to register user device for notifications: ${
-          error instanceof Error ? error.message : 'An unexpected error occurred'
-        }`
-      )
-    })
-}
+
 
 export const isDeviceAlreadyRegisteredForUser = async(
   deviceToken: string,
@@ -42,28 +37,6 @@ export const isDeviceAlreadyRegisteredForUser = async(
     registeredDevice.userId === userId
 }
 
-export const withRegisterPushOnFocus = <T>(component: T): T => {
-  const fetchClient = useFetchClient()
 
-  useFocusEffect(
-    useCallback(() => {
-      const register = async() => {
-        try {
-          registerUserDeviceForNotification(fetchClient)
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.error('Failed to register device:', error)
-        }
-      }
-      const task = InteractionManager.runAfterInteractions(() => {
-        void register()
-      })
-
-      return () => task.cancel()
-    }, [fetchClient])
-  )
-
-  return component
-}
 
 export default registerUserDeviceForNotification
